fix: 나이 소수값으로 입력하면 에러 잘 뜨도록

+++ week1/validation/index.js
@@ -35,7 +35,7 @@ const validateEmail = () => {
 // 나이 유효성 검사 함수
 const validateAge = () => {
   const age = document.getElementById("age");
-  const ageValue = parseInt(age.value);
+  const ageValue = parseFloat(age.value);
   const ageMessage = document.getElementById("ageMessage");
 
   if (age.value == "") {
